init: experimental yosys-{bluespec, ghdl} plugins
These plugins can be included in a closure, along with the `yosys`
derivation, and they will be automatically picked up for use. For
example, this allows you to include 'yosys-bluespec' in your
`buildInputs`, and then immediately run:

    $ nix-shell -p yosys yosys-bluespec yosys-ghdl
    $ yosys -m bluespec -p 'help read_bluespec'
    $ yosys -m ghdl -p 'help ghdl'

These two plugins are a bit experimental, admittedly, but they are good,
clean examples of how to write and use the yosys plugin infrastructure,
and make it easy to test updates, etc.

Signed-off-by: Austin Seipp <aseipp@pobox.com>