jmeter: 5.1.1 > 5.4
Update Jmeter to the latest version, also add myself as a maintainer

I have been using the package for 5 months daily, at the moment I am aware of 3 bugs.

1. Jmeter fails to save test plan in /nix/store (by default it saves it where Jmeter was installed).

2. The test fails to run on the latest version complaining about some missing libraries, for example log4j.

3. Jmeter fails to generate the report from a JTL file, for some reason it is not copying the needed files sometimes it fails without any log or error.

The rest works perfectly, but I still plan to fix it at somepoint, as I only use it for development, I didn't find any other bugs, and in the case of the report generator it is possible to generate it during the test run (in my case, the test is run on a non-NIXOS machine) or in a container.

Also add why tests are disabled.

Signed-off-by: Bryan A. S <bryanasdev000@gmail.com>