libvirt: don't use iptables-nftables
Per a comment on the PR that made this change, it turns out to cause
issues in some cases: https://github.com/NixOS/nixpkgs/pull/109332#issuecomment-762005163

For now, let's revert back. Presumably the issues derive from the system
iptables not matching libvirt's iptables.

In the future, #81172 should move us back into the future, and I'm
perfectly fine waiting for that PR to handle this separately.