nixos/libinput: change mouse device defaults
This commit changes some mouse device defaults that doesn't make sense
(generally) for a mouse, like tapping/tappingDragLock or
middleEmulation (it works in mouses, but unless you have a mouse with
only 3 buttons it doesn't make much sense to have it enabled). They're
now set to false or null depending of the case.

Since those options can still be useful in some exotic cases, instead of
disabling those options in mouse, I set them with `visible = false;`.
This way, they don't show in manual, but in the case the user needs it
can still be set.