ocamlPackages.ocaml_extlib: 1.7.7 -> 1.7.8
1.7.8 changed the behavior of the minimal build type (which we are
keeping as the default because opam-repository does it as well): It now
excludes the Base64 module which is prone to namespacing problems.

Since google-drive-ocamlfuse still uses the Base64 module, we need to
override it to use extlib without the minimal build type. 1.7.9 (?)
should make this obsolete as it is planned to split the Base64 module
into a separate package.

Co-authored-by: sternenseemann <sternenseemann@systemli.org>