nvidia-x11: ignore linux-rt sanity check
Ignore the sanity check that prevents the Nvidia drivers from
being built for kernels with real-time patches.
Even though the driver might not be officially supported by
Nvidia for linux-rt, it seems to work without issues.
Attempting to install the driver for linux-rt fails with the
following error message:

The kernel you are installing for is a PREEMPT_RT kernel!

The NVIDIA driver does not support real-time kernels. If you
are using a stock distribution kernel, please install
a variant of this kernel that does not have the PREEMPT_RT
patch set applied; if this is a custom kernel, please
install a standard Linux kernel.  Then try installing the
NVIDIA kernel module again.

*** Failed PREEMPT_RT sanity check. Bailing out! ***