autoreconfHook: make overridable
Some packages have specific autotools requirements that don't apply to
any other packages.  We want to be able to use autoreconfHook for
these packages with the required autotools versions, but we don't want
to have to makeSetupHook for each one, or have a top-level attribute
for every combination.

So, use callPackage around the makeSetupHook call so that the
autotools used by autoreconfHook can be overridden.  This way, a
custom autoreconfHook can be passed to a package like this:

    autoreconfHook = with buildPackages;
      autoreconfHook.override { automake = automake115x; };

And we can simplify the definitions of our existing autoreconfHook264
and autoreconfHook269 attributes.