spotify: use default libgcrypt / libpng
Instead of overriding `libgcrypt` with the (insecure) `libgcrypt_1_5`
and `libpng` with `libpng12`, use the defaults for those two packages.

Built and checked on my NixOS desktop machine because @dotlambda said he
wouldn't check proprietary software in #111215.

`spotify` was changed to use `libgcrypt_1_5` instead of `libgcrypt` in
commit 165cb05ea5f42866a5b38521a8266f10045d48b2 by @monocell in PR #8157
to address #8156, which found that:

> the current spotify client seems to depend on `libgcrypt.so.11`.
> Pretending with libgcrypt.so.20 produces an error like:
>     ... libgcrypt.so.11: version `GCRYPT_1.2' not found ...

Given that the relevant Spotify client is more than 5 years old, I don't
think we have to worry about this any more. :)

This commit helps #106203, but doesn't close it because `libgcrypt_1_5`
is still used in `staruml`.