libschrift: fix build on darwin
On darwin unsigned long int != uint_fast32_t:
https://opensource.apple.com/source/Libc/Libc-1439.40.11/include/_types/_uint32_t.h.auto.html

Thus the build fails as SFT_Glyph is defined as unsigned long, but
schrift.c uses uint_fast32_t everywhere instead of SFT_Glyph. We fix
this by adjusting the int type in the header which should be fine for
now.

This patch is technically not required for Linux, but it doesn't hurt
either as unsigned long == uint_fast32_t.