maven: use latest jdk
jdk8 is inconveniently old. Picking between `jdk` and `jdk11` is not
straightforward, but since https://maven.apache.org/download.cgi
does not list an upper bound on the JDK version I think it makes sense
to go with the latest release rather than with the latest LTS release.
Projects that want to target a specific JDK should set up their build
to achieve that anyway (or can fall back to a previous JDK if that is
hard).