fscryptctl-experimental: 2017-10-23 -> 0.1.0
Version 0.1.0 is based on the last commit that included support for V1
encryption policies. Version 1.0 is about to be released and will
include a PR which removes V1 policy support and adds V2 policy support.
Source: https://github.com/google/fscryptctl/issues/12#issuecomment-772888154

When version 1.0 is released we'll likely package it as fscryptctl and
mark fscryptctl-experimental as broken (referring to fscryptctl).