nixos/i2c: add module to set up i2c permissions
This is a very simple module that installs a single udev rule.
The rule set the ownership of all /dev/i2c-* devices to a
group, "i2c" by default but can be changed. The "uaccess" tag
also makes systemd add an ACL for users with a seat[1].

Fix issue #91771

[1]: https://enotty.pipebreaker.pl/2012/05/23/linux-automatic-user-acl-management/