firefox: switch to rust 1.49
Again, firefox-esr doesn't even get rebuilt.

This version switching is a bit inconsistent with the way we do it
for NSS, but somehow... it felt better (?) to write this way.
Anyway, I don't mind if someone improves it; I mainly wanted to
fix Firefox quickly with minimal risk of breaking other stuff.