terraform-providers: fix impurity about Nix version used
I wasn't able to figure out for which unstable version of Nix this is
supposed to work so I added nix itself to the package list. Nix will now
be the same stable version that is also provided in this nixpkgs
checkout.

To make it work with a stable Nix version I also had to trim the build
error output as otherwise it would contain the sha256 prefix & a newline
followed by "error:".