python3Packages.py-multihash: mark not broken
This was marked broken in https://github.com/NixOS/nixpkgs/pull/101636, which was merged into the staging branch. Later, the package was updated and fixed in https://github.com/NixOS/nixpkgs/pull/101801. When the staging branch was then merged into master, the package was marked as broken while actually being fixed.