mandoc: set HOMEBREWDIR on macOS
makewhatis(1) from mandoc does some validation of man pages before
indexing them. Among other things it checks if the realpath of the
man page is below the directory that is to be indexed.

For software installed with homebrew man pages under /usr/local/share/man
will be symlinked to /usr/local/Cellar making makewhatis skipping them
while indexing. Setting HOMEBREWDIR creates an exception for this case.

Setting this in nixpkgs

* doesn't hurt in any conceivable way
* accounts for the rare use case that somebody uses homebrew and nix on
  macOS and wants to index their man pages using makewhatis from
  nixpkgs.

As a side note: Yes HOMEBREWDIR could also be (ab)used to allow for
makewhatis indexing man pages on NixOS (which symlink from
/run/current-system/sw/share/man to somewhere in /nix/store). This could
then be used to write an alternative module to documentation.man which
uses mandoc instead of man-db.

I'm very interested in this and currently trying to get a proper patch
for the /nix/store situation into mandoc (ideally for Guix' /gnu/store
as well). If that works out I'm planning to contribute a module for
mandoc based man pages in NixOS.