nixos/docker: increase inotify user instances
Per issue #36214, it's really common to run into limits here with the
kernel default of '128'.

One easy way to hit this limit is to run ~130 containers, and then run
software that tails their logs in realtime (such as to aggregate them),
since each call to 'docker logs' uses an inotify user instance.

This increases it to a value recommended in an upstream issue.