json-glib: 1.4.4 → 1.6.2
- https://gitlab.gnome.org/GNOME/json-glib/-/tags/1.5.2
- https://gitlab.gnome.org/GNOME/json-glib/-/tags/1.6.0
- https://gitlab.gnome.org/GNOME/json-glib/-/tags/1.6.2

gtk-doc is enabled by default so I had to add it plus DocBook files to dependencies.

Also added GNOME team to maintainers so that it is not forgotten the next time (lethalman is no longer maintaining GNOME packages).

And correct license and clean up the code a bit while at it.