buildRustPackage: drop name from cargoDeps inputs
Using the requesting package's name as an input to fetchCargoTarball()
changes the resulting path of `cargoDeps` when version numbers change,
which in turn leads to a mismatched `cargoSha256` checksum.  Dropping
the name from the list of inputs ensures that differing versions of the
same package will use the same `cargoDeps`, and this is useful as it
preserves the ability to change or override the version number for a
package build without requiring a change of checksum.