home-assistant: reduce to two xdist job runners
This more closely replicates home-assistants upstream test environment
which runs on GitHub Actions, which according to its specifications¹
provides a 2-core CPU and `-n auto` would scale with the locally
available thread count.

This is another possible reason for the amount of race conditions
(`RuntimeError: Event loop is closed`) we were seeing, since moving to
parallel testing using xdist.

Reenable a few tests, that are likely to be caused by those race
conditions.

[1] https://docs.github.com/en/actions/reference/specifications-for-github-hosted-runners