nixos/tests/ceph-single-node-bluestore: init
The current Ceph tests use the old method for OSDs to store data on
disks, known as Filestore. This means there are no tests for the
Bluestore functionality that run on install, which means that things
like RocksDB being broken can slip through and break the Bluestore
functionality in a subtle and difficult to debug manner.

Add a test to check that Bluestore works, at least on a single node.