ripgrep: link PCRE2 dynamically
The pcre2 crate which is used by ripgrep uses pkg-config to find the
path of the PCRE2 dynamic library. If the pkg-config or the library is
not found, PCRE2 will be built/linked statically.

This change adds pkg-config to the nativeBuildInputs of ripgrep, so
that the PCRE2 library is detected and dynamically linked.