misc/vscode-extensions: improve scripts for updating extensions
The current update_installed_exts.sh always assumes that you're trying
update all installed extensions. That makes it hard to use the
get_vsixpkg function in a different setting.

This commit removes listing extensions, since it's trivial for a user to
run `code --list-extensions` but keeps and modifies the update script,
so it's more flexible.

Documentation was added to explain how the script works and explain VSC
in Nix(OS) in general.

The actual code is mostly the same.