smartmontools: drop mailutils dependency by default
mailutils depends on OpenSSL 1.0.2, which is marked as insecure and thus
prevents this package from being built on Hydra. Excluding it by default
should re-enable binary builds for users who don't need the email
feature.

This change is unlikely to cause a major disruption because there's an
existing NixOS module for smartd that handles email notifications. It
works without this package being dependant on mailutils. Since mailutils
isn't used for any other purpose besides providing email notification
for smartd, it's possible to drop it without feature loss.

This change also makes the dependance on inetutils optional. inetutils
is needed for the 'hostname' command, and is only used for providing
email support.