linuxPackages.nvidia_x11: fix sporadic kmalloc crash
Patch from Balbir Singh on the NVIDIA developer forums:

  https://forums.developer.nvidia.com/t/455-23-04-page-allocation-failure-in-kernel-module-at-random-points/155250/79

Fixes the issue described at the top of the thread, i.e. intermittent
crashes (kernel BUG) in nvkms_alloc, by replacing the custom logic for switching
between kmalloc() and vmalloc() with direct calls to kvmalloc() on recent
kernels, and similarly for kfree() and vfree() and kvfree().

On older kernels where kvmalloc() and friends are not available, the existing
custom logic is tweaked to fall back to vmalloc() for allocations larger than
PAGE_SIZE, rather than larger than 128k.  But this isn't relevant to NixOS,
since our kernel is already new enough to support kvmalloc() and friends.