python3Packages.grpcio: remove cctools from build inputs
This fixes `aarch64-darwin` build in #105026. Without this the `strip` in `PATH`
is the one from `darwin.cctools`, which doesn't preserve code signatures,
resulting in code that cannot run and pass `pythonImportsCheckPhase`.

The dependency on cctools was introduced in #44902 in 2018,
but it doesn't seem to be necessary on `x86_64-darwin` today.